Restructure directories.
