+++ Makefile
@@ -1,0 +1,3 @@
+GO=go
+all:
+	"$(GO)" build -o lmap cmd/lmap/main.go
