+++ pkg/lmap/common.go
@@ -1,0 +1,9 @@
+package lmap
+
+type ICMP struct {
+	Type        uint8
+	Code        uint8
+	Checksum    uint16
+	Identifier  uint16
+	SequenceNum uint16
+}
