+++ pkg/lmap/parse_subnet.go
@@ -1,0 +1,38 @@
+package lmap
+
+import "net"
+
+func GetAllIPsFromCIDR(cidr string) ([]net.IP, error) {
+	ip, ipnet, err := net.ParseCIDR(cidr)
+	if err != nil {
+		return nil, err
+	}
+
+	var ips []net.IP
+	for ip := ip.Mask(ipnet.Mask); ipnet.Contains(ip); inc(ip) {
+		ips = append(ips, dupIP(ip))
+	}
+	if len(ips) <= 2 {
+		return ips, nil
+	}
+	return ips[1 : len(ips)-1], nil
+}
+
+func inc(ip net.IP) {
+	for j := len(ip) - 1; j >= 0; j-- {
+		ip[j]++
+		if ip[j] > 0 {
+			break
+		}
+	}
+}
+
+func dupIP(ip net.IP) net.IP {
+	// To save space, try and only use 4 bytes
+	if x := ip.To4(); x != nil {
+		ip = x
+	}
+	dup := make(net.IP, len(ip))
+	copy(dup, ip)
+	return dup
+}
