+++ pkg/lmap/ping.go
@@ -1,0 +1,63 @@
+package lmap
+
+import (
+	"bytes"
+	"encoding/binary"
+	"log"
+	"net"
+	"time"
+)
+
+func Ping(ip net.IP) bool {
+	var icmp ICMP
+	//开始填充数据包
+	icmp.Type = 8 //8->echo message  0->reply message
+
+	recvBuf := make([]byte, 32)
+	var buffer bytes.Buffer
+
+	//先在buffer中写入icmp数据报求去校验和
+	binary.Write(&buffer, binary.BigEndian, icmp)
+	icmp.Checksum = checkSum(buffer.Bytes())
+	//然后清空buffer并把求完校验和的icmp数据报写入其中准备发送
+	buffer.Reset()
+	binary.Write(&buffer, binary.BigEndian, icmp)
+
+	conn, err := net.DialTimeout("ip4:icmp", ip.String(), 1*time.Second)
+	if err != nil {
+		return false
+	}
+	_, err = conn.Write(buffer.Bytes())
+	if err != nil {
+		log.Println("conn.Write error:", err)
+		return false
+	}
+	conn.SetReadDeadline(time.Now().Add(time.Second * 2))
+	num, err := conn.Read(recvBuf)
+	if err != nil {
+		return false
+	}
+
+	conn.SetReadDeadline(time.Time{})
+
+	return string(recvBuf[0:num]) != ""
+}
+
+func checkSum(data []byte) uint16 {
+	var (
+		sum    uint32
+		length int = len(data)
+		index  int
+	)
+	for length > 1 {
+		sum += uint32(data[index])<<8 + uint32(data[index+1])
+		index += 2
+		length -= 2
+	}
+	if length > 0 {
+		sum += uint32(data[index])
+	}
+	sum += (sum >> 16)
+
+	return uint16(^sum)
+}
